Changed how resistances will work. See description for bugs.
Hand icon stays over and of the type of the interactable most recently
in range. Also interact with thing that most recently came in range.
If you press space to blink, press-release tab to inventory, let go of
space, and then close inventory, you'll stay blinking because the Blink
keyup is eaten.
To be game accurate, the door buttons should deactivate while the door
is moving.