Can't make buttons disable while door is moving, see description.
If I want buttons to deactivate while the doors are moving I'm probably
going to have to make it so doors and buttons use an interface, which
means messing up the child/parent system I'm currently using. That is
actually probably for the best but I don't want to do it now.